Update 4
Animations are "working" except for starting jump cycle it refuses to pull and does not want to play no matter what I do.

current code would have to be completely changed to revert more like the ver that John had for us. My branch is officially the broken test branch